Removed manual garbage collection after runs.
This appeared to be exposing what may be bugs in C extensions running in other threads.
gc.collect() ought to be able to be called at any time, but if a badly written C extension has its objects swept out from under it when it wasn't expecting it, it may cause a hard interpreter crash. This is what we were seeing. The most suspect C extension was whatever implements checking the last modified time of files on Windows, but I couldn't narrow it down for sure because inserting printlines modified the timing such that I was unable to reproduce the bug.

At any rate, something in filewatcher.modulewatcher.ModuleWatcher.check_and_unload was causing the hard crash, but it went away when manual garbage collection in lyse was turned off.

We will simply shrug our shoulders at this, wonder why it didn't occur before, and move on.