gtk: Fixed bug where figure manager cleanup was not done if the analysis script raised an exception
This bug could lead to hundreds (or thousands) or windows being created if a script which creates one or more figures before crashing was run multiple times before the error in the script was fixed. This is a particular problem for multishot analysis scripts as the exception does not pause analysis execution (FYI I don't think we would want it to pause analysis on multishot errors, the current behaviour in this regard is correct).

The fix is to place the figure manager cleanup code in a "finally" block of a try/except statement so that it is always run regardless of whether the script succesfully made all of the intended figures or not. Any exceptions are still raised as before, and is not handled by the new try/except/finally block.