bugfix: Fixed bug in which columns no longer present in the dataframe (due to removing shots) were not removed from the Qt model.
This entailed changing DataFrameModel.update_row() and EditColumns.update_columns()

DataFrameModel.update_row() was modified to detect when the
Qt model has columns no longer present in the dataframe (this is the part
of the code where *new* columns are detected, as well). Defunct columns in the
Qt model are deleted and a few dictionaries are updated to reflect the new
numbering of columns.

Updated the conditional emission of the signal DataFrameModel.columns_changed()
at the end of this method. Previously only emitted when new columns were added,
is now emitted when columns were deleted as well

The only slot connected to this signal is FileBox.on_columns_changed(),
which calls EditColumns.update_columns() with some of the updated dictionaries
(EditColumns is the GUI for selecting which columns to show in the FileBox).

EditColumns.update_columns() has therefore been modified to correctly handle
the fact that the new dictionaries may have had columns removed, and not just
added as was previously the case.