feature: Make lyse handle gracefully shot files that have been deleted off disk.
Whenever lyse tries to read a HDF5 file that has been deleted, it
previously would crash. Now, the shot's row
in the filebox is marked with an icon showing it as deleted off disk
and a warning is printed. No more singleshot analysis will be run on that shot
file, but its data will be kept in the dataframe until the user removes it
from the filebox, and thus still be available for multishot analysis.

The use case for this is when an experiment is cycling just to 'keep warm' or
whatever, and shots are being deleted so as not to consume hard drive space.

I removed some of the unused optional ways of calling some of lyse's
internal methods so as to decrease the number of places where
lyse reads a HDF5 file. Previously multiple code paths were allowed - one
which was a bit cleaner (reading HDF5 files at the last minute)
 and one which was a performance optimisation (read them in advance, outside the
 GUI thread, and pass the results to the GUI-related methods). Now only
the performant way of calling these methods is allowed - which is what all
the calling code was using anyway.