feature3: Added "copy to clipboard" functionality for plots. An image can be copied to the clipboard by clicking a button in the plot window, or by ctrl-c/cmd-c when the plot window has focus.
Copying can also be done programatically like:

    from lyse.analysis_subprocess import copy_figure_to_clipboard
    copy_figure_to_clipboard(fig)

however the copying will not occur immediately - only once control is returned
to the Qt mainloop. I have not put this functionality (of copying
programatically) in a more accessible place as it is not very general - it
requires a Qt mainloop to be running.