Use existing `QApplication`
Instead of creating a new one unconditionally. The splash screen already
creates a `QApplication`, so we had two existing in the interpreter.
Having multiple `QApplication`s in the same thread (possibly in the same
process?) is not allowed and leads to segfaults on `PyQt5` and Python
exceptions on `PySide2`.

Initially reported in labscript-suite/runmanager#74