Initial pass at more detailed lyse documentation.
This is basically just a copy-paste of the old pdf documentation
already included in lyse, written by Chris Billington in 2014,
as well as the basic structure to scrape the docstrings.

For now, I'm including the :undoc-members: tag to make it more obvious
which methods still need docstrings written. Once we get better coverage,
we can remove.